Fix Scala 3 Macros
The charSequenceExprToStringExpr() caused (some?) strings to be
evaluated to

'{ ... }

It turns out, the whole function is not necessary: every invocation
of ${charSequenceExprToStringExpr(message)} can be replaced with
${message.toString}, which evaluates to the desired String (e.g., "{}
failed because of {} (baz={})").

Furthermore, the return type of deconstructInterpolatedMessage() can
be changed from Expr[String] to Expr[CharSequence]. This allows
deconstructInterpolatedMessage() to drop its usage of
charSequenceExprToStringExpr().